pyproject.toml
[project]
name = "valid-open-api-spec"
version = "0.0.0"
description = "Validates that a value is a valid OpenAPI Specification."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.3.2",
    "jsonschema",
    "pytest",
    "pyyaml",
    "referencing"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]